Update EnrollButton to link to DSC for licenses
Link to the Data Sharing Consent (DSC) view for learners enrolling with
a license.
Note that this commit assumes that the user has access to the content
through their subscription and is using a license to enroll. The logic
around calling the license subsidy view will be implemented by:
https://openedx.atlassian.net/browse/ENT-2685. Additionally, this
enrollment url will need to be updated in the future to support codes
and offers.