Pull in (and slightly modify) the site header
See the listed ticket (ENT-3343) for the full discussion on the
rationale, but here is the TL;DR:
The SiteHeader from the @edx/frontend-component-site-header package
which we used to use has largely been abandoned and is no longer
actively maintained. We have decided to pull in this header to have
full control over how our header works.
The main modification (and the original goal of this work) is to use
`Link` for internal links in the header to avoid full page refreshes in
navigation.

ENT-3343