feat: MICROBA-1521 notices redirect
To support the notices plugin on platform, this adds a redirect to the course home page. If a user lands on that page and has not acknowledged a notice, the user will be redirected to notice instead of the course home. The user can back out of the notice, or acknowledge it. They will continue to see the notice every time they hit any course until it is acknowledged.