feat: Optimizely experiment for learners with a subscription license.
Hypothesis: IF we show license learners that the course truly is free
to them at the moment of enrollment, THEN they will be more likely to
enroll, BECAUSE we will be more closely tying together the concepts of
“we want you to enroll now” and “this course is free to you”.

Implementation: If the learner has a subscription license, bring the
struck out price information in the course sidebar into the enrollment
area above the fold. 50% / 50% traffic. Assess whether or not learners
are ( 1 ) more likely to enroll in a course based on receiving vs. not
receiving the treatment.