fix: block learners from accessing ExternalCourseEnrollment outside enrollment window
BEFORE this commit: It's possible (and easy) in some cases for a learner to
enroll in a course even if late enrollment feature is disabled on the
policy.

The exact conditions which allow this bug to manifest:

* The current date is within 30 days after the enrollment deadline of a course run.
* The learner has been given a link to a ExternalCourseEnrollment page (i.e. T's&C's page) for that course run.
* The policy late enrollment is disabled (including cases where late enrollment was enabled, then self-expired).

AFTER this commit: The ExternalCourseEnrollment page (i.e. T's&C's page)
will render a 404 page in the above case.

Side effects:
* Now, this view also returns 404 *ANY* time the requested course run is
  non-redeemable according to can-redeem, or the course run is
  "unavailable" due to the current date being outside the enrollment
  window.

ENT-8856