feat: create a route to handle invite keys that includes the customer slug
The existing universal link route does not contain the enterprise slug, so for unauthenticated users,
the frontend has no knowledge of which customer’s branded logistration for which to redirect.
This is typically relied upon based on the `:enterpriseSlug` route parameter in the URL,
which is missing for the existing universal link route.
This change adds a new invite key route that includes the enterprise slug (but
not under the root loader). This will enable us to produce slug-aware invite
URLs from the admin portal, that receiving learners
can use to become linked to the enterprise via the B2B-specific auth flow (which is customer-aware).
ENT-9428